added the new programs

+++ src/JAVA_DSA/CLASS/DSA_MODULE_1/DSA_Arrays_Sliding_Window_Contribution_Technique/Subarray_with_given_sum_and_length.java
@@ -1,12 +1,12 @@
 package CLASS.DSA_MODULE_1.DSA_Arrays_Sliding_Window_Contribution_Technique;
 
 public class Subarray_with_given_sum_and_length {
-    public static int bruteforce(int[]A,int B,int C) {
+    public static int bruteforce(int[]A,int B,int C){
         int n = A.length;
         int start = 0;
         int end = B-1;
-        while (end < n){
-            int sum = 0;
+        while(end < n){
+            long sum = 0;
             for(int i = start; i <= end; i++){
                 sum+=A[i];
             }
@@ -19,7 +19,7 @@ public class Subarray_with_given_sum_and_length {
     }
     public static int solve(int[]A,int B,int C){
         int n = A.length;
-        int sum = 0;
+        long sum = 0;
         for(int i = 0 ; i < B; i++){
             sum+=A[i];
         }
@@ -28,8 +28,8 @@ public class Subarray_with_given_sum_and_length {
         }
         int start = 1;
         int end = B;
-        while (end < n){
-            sum = sum - A[start-1] + A[end];
+        while(end < n){
+            sum  = sum - A[start-1] + A[end];
             if(sum==C){
                 return 1;
             }
